+++ API-Server/Happy Habits App/Forms/WeightsStatistics.cs
@@ -2,7 +2,7 @@
 
 namespace Happy_Habits_App.Forms
 {
-    public class WeightsStatistics(double averageDuration, List<string> topExercises, int totalWorkouts, double averageExercisesPerWorkout, float averageKgsPerWorkout)
+    public class WeightsStatistics(double averageDuration, List<string> topExercises, int totalWorkouts, int averageExercisesPerWorkout, float averageKgsPerWorkout)
     {
         [JsonPropertyName("averageDuration")]
         public double AverageDuration { get; set; } = averageDuration;
@@ -11,7 +11,7 @@ namespace Happy_Habits_App.Forms
         [JsonPropertyName("totalWorkouts")]
         public int TotalWorkouts { get; set; } = totalWorkouts;
         [JsonPropertyName("averageExercisePerWorkout")]
-        public double AverageExercisesPerWorkout { get; set; } = averageExercisesPerWorkout;
+        public int AverageExercisesPerWorkout { get; set; } = averageExercisesPerWorkout;
         [JsonPropertyName("averageKgsPerWorkout")]
         public float AverageKgsPerWorkout { get; set; } = averageKgsPerWorkout;
     }
